pyproject.toml
[tool.poetry]
name = "fireside-bot"
version = "0.0.1"
description = "A bot for the Fireside Politics Discord server."
authors = ["penloy", "ilevn"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
logbook = "^1.5.3"
"discord.py" = "^1.3.1"
click = "^7.0"
uvloop = "^0.14.0"
asyncpg = "^0.20.1"
parsedatetime = "^2.5"
python-dateutil = "^2.8.1"
lru_dict = "^1.1.6"
sentry_sdk = "^0.14.2"
regex = "^2020.2.20"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
